+++ src/features/projects/data/useProjects.ts
@@ -1,5 +1,5 @@
 import useSWR from "swr"
-import fetcher from "@/common/fetcher"
+import fetcher from "@/common/utils/fetcher"
 import IProject from "../domain/IProject"
 
 type ProjectContainer = { projects: IProject[] }
